requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
ultralytics>=8.3.0
opencv-python-headless>=4.10.0
python-multipart>=0.0.9
pydantic>=2.9.2
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
numpy>=1.26.0
pillow>=10.4.0

# ML model dependencies (YOLOv11 + OpenCV + Whisper)
openai-whisper>=20231117
sounddevice>=0.4.6
pyyaml>=6.0
onnx>=1.15.0
onnxruntime>=1.16.0
